+++ src/webapi/static/i18n/en.json
@@ -18,10 +18,8 @@
   "common_legend_running_avg": "Running average",
   "app_add": "Add",
   "app_add_ed2k_link": "Add ed2k link",
-  "app_confirm_disconnect_both": "Disconnect from ED2K and Kad?",
   "app_connected": "Connected",
   "app_connecting": "Connecting…",
-  "app_disconnect": "Disconnect",
   "app_disconnected": "Disconnected",
   "app_ed2k": "ED2K",
   "app_error": "Error",
@@ -97,8 +95,6 @@
   "networks_server_confirm_remove": "Remove server “{name}”?",
   "networks_server_connect": "Connect",
   "networks_server_description": "Description",
-  "networks_server_disconnect": "Disconnect ED2K",
-  "networks_server_disconnected": "Disconnected",
   "networks_server_empty": "Nothing here yet.",
   "networks_server_error": "Error",
   "networks_server_files": "Files",
@@ -118,9 +114,7 @@
   "networks_server_version": "Version",
   "networks_kad_bootstrap_from_node": "Bootstrap from node",
   "networks_kad_buddy": "Buddy",
-  "networks_kad_connect": "Connect from known clients",
   "networks_kad_contacts_nodes": "Contacts (nodes)",
-  "networks_kad_disconnect": "Disconnect Kad",
   "networks_kad_error": "Error",
   "networks_kad_files": "Files",
   "networks_kad_firewalled_tcp": "Firewalled (TCP)",
@@ -135,8 +129,6 @@
   "networks_kad_no": "No",
   "networks_kad_state": "State",
   "networks_kad_toast_bootstrapping": "Bootstrapping…",
-  "networks_kad_toast_connecting": "Connecting Kad…",
-  "networks_kad_toast_disconnecting": "Disconnecting Kad…",
   "networks_kad_toast_enter_ip_port": "Enter IP:port",
   "networks_kad_users": "Users",
   "networks_kad_yes": "Yes",
@@ -582,6 +574,7 @@
   "downloads_peer_state_error": "Error",
   "networks_kad_conn_connected": "Connected",
   "networks_kad_conn_connecting": "Connecting…",
+  "networks_kad_conn_disabled": "Disabled",
   "networks_kad_conn_disconnected": "Disconnected",
   "prefs_general": "General",
   "prefs_connection": "Connection",
