+++ src/webapi/static/i18n/es.json
@@ -18,10 +18,8 @@
   "common_legend_running_avg": "Media móvil",
   "app_add": "Añadir",
   "app_add_ed2k_link": "Añadir enlace ed2k",
-  "app_confirm_disconnect_both": "¿Desconectar de ED2K y Kad?",
   "app_connected": "Conectado",
   "app_connecting": "Conectando…",
-  "app_disconnect": "Desconectar",
   "app_disconnected": "Desconectado",
   "app_ed2k": "ED2K",
   "app_error": "Error",
@@ -97,8 +95,6 @@
   "networks_server_confirm_remove": "¿Eliminar el servidor «{name}»?",
   "networks_server_connect": "Conectar",
   "networks_server_description": "Descripción",
-  "networks_server_disconnect": "Desconectar ED2K",
-  "networks_server_disconnected": "Desconectado",
   "networks_server_empty": "Aún no hay nada aquí.",
   "networks_server_error": "Error",
   "networks_server_files": "Archivos",
@@ -118,9 +114,7 @@
   "networks_server_version": "Versión",
   "networks_kad_bootstrap_from_node": "Bootstrap desde nodo",
   "networks_kad_buddy": "Compañero",
-  "networks_kad_connect": "Conectar desde clientes conocidos",
   "networks_kad_contacts_nodes": "Contactos (nodos)",
-  "networks_kad_disconnect": "Desconectar Kad",
   "networks_kad_error": "Error",
   "networks_kad_files": "Archivos",
   "networks_kad_firewalled_tcp": "Tras cortafuegos (TCP)",
@@ -135,8 +129,6 @@
   "networks_kad_no": "No",
   "networks_kad_state": "Estado",
   "networks_kad_toast_bootstrapping": "Arrancando (bootstrap)…",
-  "networks_kad_toast_connecting": "Conectando Kad…",
-  "networks_kad_toast_disconnecting": "Desconectando Kad…",
   "networks_kad_toast_enter_ip_port": "Introduce IP:puerto",
   "networks_kad_users": "Usuarios",
   "networks_kad_yes": "Sí",
@@ -582,6 +574,7 @@
   "downloads_peer_state_error": "Error",
   "networks_kad_conn_connected": "Conectado",
   "networks_kad_conn_connecting": "Conectando…",
+  "networks_kad_conn_disabled": "Desactivado",
   "networks_kad_conn_disconnected": "Desconectado",
   "prefs_general": "General",
   "prefs_connection": "Conexión",
